factor Track component out of BlockBasedTrack
